Consolidate reusable functions in github-actions to utils folder (#5944)
* consolidate reusable functions in github-actions to utils folder

* add global variables to utils functions arguments

* update utils/format-comment.js function definition

* remove variable fs from object destructuring in utils/format-comment.js function argument
