+++ github-actions/trigger-schedule/add-update-label-weekly/add-label.js
@@ -1,5 +1,6 @@
 // Import modules
 const findLinkedIssue = require('../../utils/find-linked-issue');
+const getTimeline = require('../../utils/get-timeline');
 var fs = require("fs");
 // Global variables
 var github;
@@ -32,7 +33,7 @@ async function main({ g, c }, columnId) {
   // Retrieve all issue numbers from a column
   const issueNums = getIssueNumsFromColumn(columnId);
   for await (let issueNum of issueNums) {
-    const timeline = await getTimeline(issueNum);
+    const timeline = await getTimeline(issueNum, github, context);
     const timelineArray = Array.from(timeline);
     const assignees = await getAssignees(issueNum);
     // Error catching.
@@ -95,39 +96,6 @@ async function* getIssueNumsFromColumn(columnId) {
       page++;
     }
   }
-}
-/**
- * Function that returns the timeline of an issue.
- * @param {Number} issueNum the issue's number
- * @returns an Array of Objects containing the issue's timeline of events
- */
-
-async function getTimeline(issueNum) {
-  let arra = []
-  let page = 1
-  while (true) {
-    try {
-      const results = await github.rest.issues.listEventsForTimeline({
-        owner: context.repo.owner,
-        repo: context.repo.repo,
-        issue_number: issueNum,
-        per_page: 100,
-        page: page,
-      });
-      if (results.data.length) {
-        arra = arra.concat(results.data);
-      } else {
-        break
-      }
-    } catch (err) {
-      console.log(error);
-      continue
-    }
-    finally {
-      page++
-    }
-  }
-  return arra
 }
 
 /**
